+++ land_consumption/calculation.py
@@ -4,10 +4,11 @@ from geopandas import GeoDataFrame
 from land_consumption.utils import LandObjectCategory, SQM_TO_HA_FACTOR, LandUseCategory
 
 
-def calculate_land_consumption(aoi_area: float, area_df: GeoDataFrame) -> pd.DataFrame:
+def calculate_land_consumption(area_df: GeoDataFrame) -> pd.DataFrame:
     area_df['Total Land Area [ha]'] = area_df['area'] * SQM_TO_HA_FACTOR
-    aoi_area_ha = aoi_area * SQM_TO_HA_FACTOR
-    area_df['% of Settled Land Area'] = (area_df['Total Land Area [ha]'] / aoi_area_ha) * 100
+
+    settled_land = area_df.apply(lambda x: x['area'] if is_land_settled(x) else None, axis='columns')
+    area_df['% of Settled Land Area'] = settled_land / settled_land.sum() * 100
 
     consumed_land = area_df.apply(lambda x: x['area'] if is_land_consumed(x) else None, axis='columns')
     area_df['% of Consumed Land Area'] = consumed_land / consumed_land.sum() * 100
@@ -15,9 +16,18 @@ def calculate_land_consumption(aoi_area: float, area_df: GeoDataFrame) -> pd.Dat
     area_df['Land Use Object'] = area_df['category'].apply(lambda x: x.value)
     area_df['Land Use Class'] = area_df['landuse_category'].apply(lambda x: x.value)
 
-    agri_mask = (area_df['Land Use Object'] == 'Built up land') & (area_df['Land Use Class'] == 'Agricultural')
-    area_df.loc[agri_mask, 'Land Use Object'] = 'Agricultural land'
-    area_df.loc[agri_mask, 'Land Use Class'] = ''
+    mask = (area_df['Land Use Object'] == 'Built up land') & (
+        area_df['Land Use Class'].isin(['Agricultural', 'Natural'])
+    )
+
+    area_df.loc[mask, 'Land Use Object'] = area_df.loc[mask, 'Land Use Class'].map(
+        {
+            'Agricultural': 'Agricultural land',
+            'Natural': 'Natural land',
+        }
+    )
+
+    area_df.loc[mask, 'Land Use Class'] = ''
 
     return area_df[
         [
@@ -33,8 +43,18 @@ def calculate_land_consumption(aoi_area: float, area_df: GeoDataFrame) -> pd.Dat
 def is_land_consumed(feature: pd.Series) -> bool:
     if feature['category'] == LandObjectCategory.OTHER:
         return False
-    elif (feature['category'] == LandObjectCategory.BUILT_UP) & (
-        feature['landuse_category'] == LandUseCategory.AGRICULTURAL
+    elif (feature['category'] == LandObjectCategory.BUILT_UP) and (
+        feature['landuse_category'] in [LandUseCategory.AGRICULTURAL, LandUseCategory.NATURAL]
+    ):
+        return False
+
+    return True
+
+
+def is_land_settled(feature: pd.Series) -> bool:
+    # TODO the naming here makes this line needlessly confusing fix this in review
+    if (feature['category'] == LandObjectCategory.BUILT_UP) and (
+        feature['landuse_category'] == LandUseCategory.NATURAL
     ):
         return False
     return True